refat(home-dev): create a separate component for the dev section
